+++ libpkg/pkg_format.c
@@ -1,0 +1,165 @@
+/*-
+ * Copyright (c) 2017 Vsevolod Stakhov <vsevolod@FreeBSD.org>
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer
+ *    in this position and unchanged.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE AUTHOR(S) ``AS IS'' AND ANY EXPRESS OR
+ * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+ * IN NO EVENT SHALL THE AUTHOR(S) BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
+ * NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
+ * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#include <archive.h>
+#include <archive_entry.h>
+#include <assert.h>
+#include <errno.h>
+#include <fcntl.h>
+#include <string.h>
+#include <stdlib.h>
+
+#include "pkg.h"
+#include "private/event.h"
+#include "private/pkg.h"
+
+int
+pkg_open_format(struct pkg **pkg_p, struct archive **a, struct archive_entry **ae,
+		const char *path, struct pkg_manifest_key *keys, int flags, int fd)
+{
+	struct pkg	*pkg = NULL;
+	pkg_error_t	 retcode = EPKG_OK;
+	int		 ret;
+	const char	*fpath;
+	bool		 manifest = false;
+	bool		 read_from_stdin = 0;
+
+	*a = archive_read_new();
+	archive_read_support_filter_all(*a);
+	archive_read_support_format_tar(*a);
+
+	/* archive_read_open_filename() treats a path of NULL as
+	 * meaning "read from stdin," but we want this behaviour if
+	 * path is exactly "-". In the unlikely event of wanting to
+	 * read an on-disk file called "-", just say "./-" or some
+	 * other leading path. */
+
+	if (fd == -1) {
+		read_from_stdin = (strncmp(path, "-", 2) == 0);
+
+		if (archive_read_open_filename(*a,
+				read_from_stdin ? NULL : path, 4096) != ARCHIVE_OK) {
+			if ((flags & PKG_OPEN_TRY) == 0)
+				pkg_emit_error("archive_read_open_filename(%s): %s", path,
+						archive_error_string(*a));
+
+			retcode = EPKG_FATAL;
+			goto cleanup;
+		}
+	} else {
+		if (archive_read_open_fd(*a, fd, 4096) != ARCHIVE_OK) {
+			if ((flags & PKG_OPEN_TRY) == 0)
+				pkg_emit_error("archive_read_open_fd: %s",
+						archive_error_string(*a));
+
+			retcode = EPKG_FATAL;
+			goto cleanup;
+		}
+	}
+
+	retcode = pkg_new(pkg_p, PKG_FILE);
+	if (retcode != EPKG_OK)
+		goto cleanup;
+
+	pkg = *pkg_p;
+
+	while ((ret = archive_read_next_header(*a, ae)) == ARCHIVE_OK) {
+		fpath = archive_entry_pathname(*ae);
+		if (fpath[0] != '+')
+			break;
+
+		if (!manifest &&
+				(flags & PKG_OPEN_MANIFEST_COMPACT) &&
+				strcmp(fpath, "+COMPACT_MANIFEST") == 0) {
+			char *buffer;
+			manifest = true;
+
+			size_t len = archive_entry_size(*ae);
+			buffer = xmalloc(len);
+			archive_read_data(*a, buffer, archive_entry_size(*ae));
+			ret = pkg_parse_manifest(pkg, buffer, len, keys);
+			free(buffer);
+			if (ret != EPKG_OK) {
+				retcode = EPKG_FATAL;
+				goto cleanup;
+			}
+			/* Do not read anything more */
+			break;
+		}
+		if (!manifest && strcmp(fpath, "+MANIFEST") == 0) {
+			manifest = true;
+			char *buffer;
+
+			size_t len = archive_entry_size(*ae);
+			buffer = xmalloc(len);
+			archive_read_data(*a, buffer, archive_entry_size(*ae));
+			ret = pkg_parse_manifest(pkg, buffer, len, keys);
+			free(buffer);
+			if (ret != EPKG_OK) {
+				if ((flags & PKG_OPEN_TRY) == 0)
+					pkg_emit_error("%s is not a valid package: "
+							"Invalid manifest", path);
+
+				retcode = EPKG_FATAL;
+				goto cleanup;
+			}
+
+			if (flags & PKG_OPEN_MANIFEST_ONLY)
+				break;
+		}
+	}
+
+	if (ret != ARCHIVE_OK && ret != ARCHIVE_EOF) {
+		if ((flags & PKG_OPEN_TRY) == 0)
+			pkg_emit_error("archive_read_next_header(): %s",
+					archive_error_string(*a));
+
+		retcode = EPKG_FATAL;
+	}
+
+	if (ret == ARCHIVE_EOF)
+		retcode = EPKG_END;
+
+	if (!manifest) {
+		retcode = EPKG_FATAL;
+		if ((flags & PKG_OPEN_TRY) == 0)
+			pkg_emit_error("%s is not a valid package: no manifest found", path);
+	}
+
+	cleanup:
+	if (retcode != EPKG_OK && retcode != EPKG_END) {
+		if (*a != NULL) {
+			archive_read_close(*a);
+			archive_read_free(*a);
+		}
+		free(pkg);
+		*pkg_p = NULL;
+		*a = NULL;
+		*ae = NULL;
+	}
+
+	return (retcode);
+}
